requirements.txt
Django==4.2.7
django-environ==0.12.0
django-ratelimit==4.1.0
argon2-cffi==25.1.0
django-recaptcha==3.0.0
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2024.1
